Add action hook before sending customer invoice
This is the only notification type that can't be hooked before starting
to compile the message.

One good reason for hooking there is to switch e-mail languages
according to a stored order language. Other people might find other
uses for these hooks, but, bottom line is, no hook exists in this case.