Use delegated event handler to delete tax rows
When a tax row is dynamically inserted, it can not be dynamically
deleted until the page is refreshed (becasue the event handler is
bound to only on page load). This commit delegates the event handler
from an element which is guaranteed to exist on page load, so that
dynamically inserted tax rows can also be dynamically deleted.