Filter Request: Login Credentials
I'm trying to add a "remember me" option to my login form. WordPress
supports this, but it's hardcoded to always on in the WC form handler.
Since you can't tell a difference in PHP between an unchecked checkbox
and no checkbox at all, this seems like the simplest alternative that
maintains backwards compatibility. Thanks.