Fix completing order payment for deleted product
When marking payment as complete on an order which contains a product which has
since been deleted, a fatal error will be thrown because the product is assumed
to exist.

The specific error is `PHP Fatal error: Call to a member function is_downloadable()`.
It occurs when checking if the order status should be set as completed or processing,
by determining if the product on the order is downloadable and virtual.

This patch will mark an order as "processing" if the product has since been deleted.