Update class-wc-meta-box-coupon-data.php
Introduce `woocommerce_coupon_data_panels` action.

If someone wants to add a new tab to the Coupons data meta box, they can use the `woocommerce_coupon_data_tabs` filter to register a new tab. These tabs accept a `target` key that lets you specify the `div` ID to reveal when the tab is clicked on. It turns out, however, that there is no way to actually add the `div` to the page, making the filter for custom tabs only good for removing tabs.