Display order & item totals with correct currency
If a store changes currency or is using an extension which allows for
purchasing in multiple currencies, the currency symbol displayed on the
Edit Order admin screen may be incorrect for previous orders because
it uses the `wc_price()` fallback, which is the site wide currency
rather than the currency for that specific order. This fixes that.