Rather than include all files on construct, include files after plugins_loaded
This ensures functions intended to be pluggable actually are, and
prevents race conditions. Seen some authors mess with the plugin load
order to work around this and it’s horrible.

This hook has priority 5 to ensure existing plugins using WC
functionality on plugins_loaded do not break.

cc @claudiosmweb @barrykooij