Register shop_order post statuses earlier
Registering the post statuses on `init` with a priority of 10 is too
late for some uses. Particularly if you are using `WP_Query` on `init`,
WP core will not add the `post_status` where clause if the custom post
statuses are not registered yet.

I ran into this issue on a site using `ALTERNATE_WP_CRON` which runs
`wp_cron()` on `init` with a default priority of `10`. It appears that
this may fix #7408 as well.