Compound tax calculation test
There is either an error in the calculation for compound taxes, or
there is something wrong with my understanding of how compound taxes
should work - I’m definitely not discounting the later.

Given a price of $9.99 exclusive of tax, applying a 10% tax will give
9.99 + 0.999 = 10.989. Then applying a compound tax of 2% should give
10.989 + 0.2198 = 11.2088.

The result of the WC_Tax::calc_tax on the 2% compound tax is 0.1998
which seems to be in error.

Likewise for prices exclusive of tax.

Note: rounding to 4 decimal places on all calculations