Fixes issue with adding empty order item meta
The changes just check to make sure that either a meta key or meta
value has been entered before it saves. Otherwise it deletes the order
item meta row that has been added. The user must add something to the
order item meta row now or it won’t get saved.

Not sure if this is the cleanest way to fix this issue. I was expecting
to see the function wc_update_order_item_meta() but it is not used in
this case.