Allow trashed pages to be republished
Expanding on @mehulkaklotar’s fix above. One issue is that is a deleted
store page has had its short code removed then it won’t be republished
and that page is still missing.

This fix catches pages that are in the trash but have no short code.
Those pages are now left in the trash and new page with the correct
short code is added in its place.

Fixes #8269