Add email notification for partial refunds
fixes #8278

Emails are now sent to the customer when a partial refund is issued. I
referenced #7785 when doing this.

When a full refund happens the email the customer receives is clear
enough that they are getting a full refund. However I thought it would
be useful to add a new line in these new partial refund emails that
states exactly the value of the partial refund as it may not be that
clear.