Optimization. Don't call `get_post_meta` 17+ odd times.
As @pmgarman remarked here:
https://pmgarman.me/blog/woocommerce-and-coupon-_used_by-logging/ — for
very large values of postmeta, when using memcached that cannot handle
data stores of over 1M, the store seems to fail at pulling it back in,
resulting in large query results on each call.

At the least, this change should avoid those cache misses and
re-queries, although I do feel that not being able to keep the cached
in php’s memory and losing it in the cache is indeed a failing here of
the object cache plugin.

Still, this should have some optimizations, as we’re caching the data
right here, closer to usage.