Removing selection priority
This option was added so you could sort methods in a custom order, but
choose a different default.

Doing this per zone would add extra complexity, and this has also
proven problematic when a method registers multiple rates.

This commit removes usage of the setting. We should use the display
order and choose the top method by default. It think this balances ease
of use with control.