Resolved issue #11428.
report.js export_csv of the 'table' data_format will not download if the cell data contains #.   Most likely due to the fact that #<id> is a octect character.

This is a quick fix that strips any # character from table cell data.

There may be a way to escape # however \# nor HTML encoding to %23 resolved the issue.