Rewrite rule improvements
Closes #11500

Extra rules are needed to support subpages, but the previous method was
very greedy and affected other URIs.

To improve this, loop over shop subpages and add rules specific to
those to avoid 404s.

Also flushes permalinks when a subpage of shop is saved so these rules
are correctly generated.