Add $load_address to 'woocommerce_my_account_edit_address_title'
Add $load_address to the 'woocommerce_my_account_edit_address_title' filter so that scripts can determine which address is being edited. Currently there is no obvious way to determine what address is being edited if it is not the 'billing' or 'shipping' built-in addresses. By adding $load_address we can use that to detect the name of the address being edited and display the corresponding title.