Utilise $product method to get thumbnail
Is there any reason this function isn't using the `get_image()` method? It seems if we're accessing the `$post->ID`, then we should have access to the `$product` object too?

The reason being, the `get_image()` method also checks for the parent featured image, so is more robust. It may also be worth considering a filter in the `get_image()` method.