Display error message when it's impossible to get a sample of the file.
When the file is empty or using an incorrect charset will display just an empty screen, since we output errors before check if the file is valid.

I included the error output inside the mapping step in order to fix this issue, and added a new option to display buttons after the error  message.

Closes #16928