Filter for line items in abstract-wc-order.php (#17759)
* Filter for line items in abstract-wc-order.php

When creating custom product class there are places where line items is only value that can be handled. This is one of that places causing subtotal values after checkout not to be calculated and displayed.

* Extension to woocommerce_order_get_items filter

This is extension to filter as get_items() is used for different classes and types. This way it would be possible to distinct which type is called for items.

* Removed filtering inside subtotal function

$types parameter is added as parameter for filter inside get_items() function so previous commit is reverted as there is no need for it.